requirements.txt
Flask
sentence-transformers
scikit-learn
plotly
google-generativeai
python-dotenv
pandas
numpy